+++ src/command_modules/azure-cli-find/azure/cli/command_modules/find/__init__.py
@@ -22,6 +22,8 @@ class FindCommandsLoader(AzCommandsLoader):
 
     def load_arguments(self, command):
         from azure.cli.command_modules.find._params import load_arguments
+        with self.argument_context('find') as c:
+            c.ignore('_subscription')  # hide global subscription param
         load_arguments(self, command)