+++ src/command_modules/azure-cli-rdbms/azure/cli/command_modules/rdbms/custom.py
@@ -146,7 +146,7 @@ def _server_georestore(cmd, client, resource_group_name, server_name, sku_name, 
 
 
 def _server_update_custom_func(instance,
-                               capacity=None,
+                               sku_name=None,
                                storage_mb=None,
                                backup_retention_days=None,
                                administrator_login_password=None,
@@ -158,16 +158,18 @@ def _server_update_custom_func(instance,
     module = import_module(server_module_path.replace('server', 'server_update_parameters'))
     ServerUpdateParameters = getattr(module, 'ServerUpdateParameters')
 
-    if capacity is not None:
-        instance.sku.name = _get_sku_name(instance.sku.tier, instance.sku.family, capacity)
-        instance.sku.capacity = capacity
+    if sku_name:
+        instance.sku.name = sku_name
+        instance.sku.capacity = None
+        instance.sku.family = None
+        instance.sku.tier = None
     else:
         instance.sku = None
 
-    if storage_mb is not None:
+    if storage_mb:
         instance.storage_profile.storage_mb = storage_mb
 
-    if backup_retention_days is not None:
+    if backup_retention_days:
         instance.storage_profile.backup_retention_days = backup_retention_days
 
     params = ServerUpdateParameters(sku=instance.sku,