+++ src/azure-cli-core/HISTORY.rst
@@ -9,7 +9,7 @@ Release History
 
 2.0.35
 ++++++
-* Minor fixes
+* Add Authorization Resource into the stack profile
 
 2.0.34
 ++++++