+++ src/azure-cli-core/HISTORY.rst
@@ -5,7 +5,7 @@ Release History
 
 2.0.37
 ++++++
-* Minor fixes
+* Add global support for `--subscription` to most commands.
 
 2.0.35
 ++++++